MAINT: reorganize tests
